requirements.txt
openai==1.97.1
openai-agents==0.2.4
tiktoken==0.4.0
azure-functions==1.15.0
azure-data-tables==12.4.3
azure-search-documents==11.4.0b8
azure-storage-blob==12.17.0
azure-identity==1.15.0
azure-core==1.28.0
pyyaml==6.0.2
fastapi==0.115.12
uvicorn==0.30.5
httpx==0.27.2
multipart==0.2.5
openpyxl==3.1.5
pandas==2.0.3
numpy==1.24.3
typing_extensions==4.13.2
thefuzz==0.20.0
azure-ai-formrecognizer==3.3.3
PyPDF2==3.0.1
Pillow==11.0.0
pdf2image==1.17.0
python-Levenshtein==0.25.1
aiohttp==3.11.11
aiosignal==1.4.0
aiofiles==24.1.0
mlflow==2.22.0
newspaper3k==0.2.8
lxml_html_clean==0.4.2
python-dotenv==1.1.0